Word things now in doubles for extra precision.

+++ src/game/celestial/Orbit.java
@@ -4,19 +4,19 @@ package game.celestial;
  * A realistic orbit object.
  */
 public class Orbit {
-	private float eccentricity; // The orbits deviation, 0<e<1.
+	private double eccentricity; // The orbits deviation, 0<e<1.
 
-	private float semiMajorAxis; // The semi-major axis (AU).
-	private float semiMinorAxis; // The semi-minor axis (AU).
+	private double semiMajorAxis; // The semi-major axis (AU).
+	private double semiMinorAxis; // The semi-minor axis (AU).
 
-	private float periapsis; // The orbits size periapsis (AU).
-	private float apoapsis; // The orbits size apoapsis (AU).
-	private float period; // The orbital period (Earth Days).
-	private float velocity; // The objects velocity (km/s).
+	private double periapsis; // The orbits size periapsis (AU).
+	private double apoapsis; // The orbits size apoapsis (AU).
+	private double period; // The orbital period (Earth Days).
+	private double velocity; // The objects velocity (km/s).
 
-	private float pitch; // The orbits inclination, orbits prograde 0<i<90, orbits retrograde 90<i<180.
-	private float yaw; // The orbits longitude of the ascending node from 0-360 degrees.
-	private float roll; // The orbits argument of periapsis from 0-360 degrees.
+	private double pitch; // The orbits inclination, orbits prograde 0<i<90, orbits retrograde 90<i<180.
+	private double yaw; // The orbits longitude of the ascending node from 0-360 degrees.
+	private double roll; // The orbits argument of periapsis from 0-360 degrees.
 
 	/**
 	 * A object that represents a realistic orbit.
@@ -28,59 +28,59 @@ public class Orbit {
 	 * @param yaw The orbits longitude of the ascending node from 0-360 degrees.
 	 * @param roll The orbits argument of periapsis from 0-360 degrees.
 	 */
-	public Orbit(float eccentricity, float semiMajorAxis, float solarMasses, float pitch, float yaw, float roll) {
+	public Orbit(double eccentricity, double semiMajorAxis, double solarMasses, double pitch, double yaw, double roll) {
 		this.eccentricity = eccentricity;
 
 		this.semiMajorAxis = semiMajorAxis;
-		this.semiMinorAxis = semiMajorAxis * (float) Math.sqrt(1 - Math.pow(eccentricity, 2.0f));
+		this.semiMinorAxis = semiMajorAxis * Math.sqrt(1.0 - Math.pow(eccentricity, 2.0));
 
-		this.periapsis = semiMajorAxis * (float) Math.sqrt(1 - eccentricity);
-		this.apoapsis = semiMajorAxis * (float) Math.sqrt(1 + eccentricity);
-		this.period = (float) Math.sqrt(Math.pow(semiMajorAxis, 3.0f) / solarMasses) * 365.25f;
-		this.velocity = (float) Math.sqrt(solarMasses / semiMajorAxis) * 29.78f;
+		this.periapsis = semiMajorAxis * Math.sqrt(1.0 - eccentricity);
+		this.apoapsis = semiMajorAxis * Math.sqrt(1.0 + eccentricity);
+		this.period = Math.sqrt(Math.pow(semiMajorAxis, 3.0) / solarMasses) * 365.25;
+		this.velocity = Math.sqrt(solarMasses / semiMajorAxis) * 29.78;
 
 		this.pitch = pitch;
 		this.yaw = yaw;
 		this.roll = roll;
 	}
 
-	public float getEccentricity() {
+	public double getEccentricity() {
 		return eccentricity;
 	}
 
-	public float getSemiMajorAxis() {
+	public double getSemiMajorAxis() {
 		return semiMajorAxis;
 	}
 
-	public float getSemiMinorAxis() {
+	public double getSemiMinorAxis() {
 		return semiMinorAxis;
 	}
 
-	public float getPeriapsis() {
+	public double getPeriapsis() {
 		return periapsis;
 	}
 
-	public float getApoapsis() {
+	public double getApoapsis() {
 		return apoapsis;
 	}
 
-	public float getPeriod() {
+	public double getPeriod() {
 		return period;
 	}
 
-	public float getVelocity() {
+	public double getVelocity() {
 		return velocity;
 	}
 
-	public float getPitch() {
+	public double getPitch() {
 		return pitch;
 	}
 
-	public float getYaw() {
+	public double getYaw() {
 		return yaw;
 	}
 
-	public float getRoll() {
+	public double getRoll() {
 		return roll;
 	}
